requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
xgboost
shap
Pillow

